# src/main/resources/openapi/pennywise-v1.yaml
openapi: 3.0.3
info:
  title: PennyWise API
  version: 1.0.0
  description: Personal expense tracking API with multi-book support, idempotent operations, and optimistic concurrency control
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /v1/books/{bookId}/categories:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [Categories]
      summary: List categories for book
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CategoryResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Categories]
      summary: Add category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryCreateRequest" }
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/books/{bookId}/categories/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/BookId"
      - $ref: "#/components/parameters/CategoryId"
      - $ref: "#/components/parameters/IfMatch"
    patch:
      tags: [Categories]
      summary: Rename/disable category (disable preferred over delete)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryUpdateRequest" }
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        example: '"3"'
      description: ETag value from prior GET (optimistic concurrency)
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionFailed:
      description: Precondition Failed (ETag mismatch)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string, example: "Invalid request" }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            requestId: { type: string, example: "01HZX..." }
      required: [error]
...
    CategoryType:
      type: string
      enum: [INCOME, EXPENSE]
...
    CategoryCreateRequest:
      type: object
      required: [type, name]
      properties:
        type: { $ref: "#/components/schemas/CategoryType" }
        name: { type: string, minLength: 1, maxLength: 60 }
        icon: { type: string, maxLength: 64, nullable: true, example: "utensils" }
        color: { type: string, pattern: "^#[0-9A-Fa-f]{6}$", nullable: true, example: "#EF4444" }
    CategoryUpdateRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        isDisabled: { type: boolean }
        icon: { type: string, maxLength: 64, nullable: true }
        color:
          type: string
          pattern: '^(?:\s*|#[0-9A-Fa-f]{6})$'
          nullable: true
          description: A blank value clears the category color.
    CategoryResponse:
      type: object
      required: [id, bookId, type, name, isDisabled, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        bookId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/CategoryType" }
        name: { type: string }
        isDisabled: { type: boolean }
        icon: { type: string, nullable: true }
        color: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        version: { type: integer, format: int64 }
...